test: add coverage for candidate utility modules
