+++ tests/test_candidate_ranking.py
@@ -7,7 +7,7 @@ ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
 sys.path.insert(0, str(ROOT / "server"))
 
-from server.steps.candidates.helpers import _enforce_non_overlap
+from server.steps.candidates.refine import _enforce_non_overlap
 from server.interfaces.clip_candidate import ClipCandidate
